Implements Cache Namespace
A cache namespace represents a group of keys holding a timestamp with
nano precision.

It's intended to be used by providing keys of groups of items to get the
most recent timestamp amongst them in order to compose a prefix of
keys in subsequent cache calls.

Once a namespace is invalidated by Delete, the current clock time will be
returned, breaking the cache automatically as long as the timestamp is
used as prefix.